pyproject.toml
[project]
name = "img-resized-512"
version = "0.1.0"
description = ""
authors = [
    {name = "King-Navi",email = "168604098+King-Navi@users.noreply.github.com"}
]
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.3.4,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
